refactor!: make KtorEngine generic over any Ktor http client engine
Deprecate KtorEngine as synonymous with OkHttp. The `http-client-engine-ktor` is
now a utility wrapper for _any_ Ktor compliant engine. The OkHttp engine
has been moved to `http-client-engine-default` as the default engine on
JVM.

BREAKING CHANGE:

Use of explicit `httpClientEngine = KtorEngine()` will no longer
compile. The default engine will be synonymous with OkHttp
making it unnecessary to set this engine. Users should remove
this explicit SDK client configuration.
