+++ runtime/protocol/http-client-engines/http-client-engine-default/build.gradle.kts
@@ -6,6 +6,8 @@
 description = "Default HTTP Client Engine for Smithy services generated by smithy-kotlin"
 extra["moduleName"] = "aws.smithy.kotlin.runtime.http.engine"
 
+val ktorVersion: String by project
+
 kotlin {
     sourceSets {
         commonMain {
@@ -16,6 +18,8 @@ kotlin {
         jvmMain {
             dependencies {
                 implementation(project(":runtime:protocol:http-client-engines:http-client-engine-ktor"))
+                // okhttp works on both JVM and Android
+                implementation("io.ktor:ktor-client-okhttp:$ktorVersion")
             }
         }
         all {
