+++ runtime/protocol/http-client-engines/http-client-engine-ktor/common/test/aws/smithy/kotlin/runtime/http/engine/ktor/KtorEngineTest.kt
@@ -13,9 +13,8 @@ import kotlinx.coroutines.test.runTest
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.time.Duration.Companion.milliseconds
-import kotlin.time.ExperimentalTime
 
-@OptIn(ExperimentalCoroutinesApi::class, ExperimentalTime::class)
+@OptIn(ExperimentalCoroutinesApi::class)
 class WaiterTest {
     @Test
     fun testSignalWhenWaiting() = runTest {
