+++ runtime/protocol/http-client-engines/http-client-engine-ktor/common/test/aws/smithy/kotlin/runtime/http/engine/ktor/KtorUtilsTest.kt
@@ -15,14 +15,11 @@ import io.ktor.http.*
 import io.ktor.http.Headers
 import io.ktor.http.HttpMethod
 import io.ktor.http.HttpStatusCode
-import io.ktor.util.*
 import io.ktor.util.date.GMTDate
 import io.ktor.utils.io.*
 import kotlin.coroutines.CoroutineContext
 import kotlin.test.*
 
-// TODO: Remove following annotation after https://youtrack.jetbrains.com/issue/KTOR-3001 is resolved
-@OptIn(InternalAPI::class)
 class MockHttpResponse : HttpResponse() {
     override val call: HttpClientCall
         get() = error("not needed for test")
@@ -38,8 +35,6 @@ class MockHttpResponse : HttpResponse() {
     override val version: HttpProtocolVersion = HttpProtocolVersion.HTTP_1_1
 }
 
-// TODO: Remove following annotation after https://youtrack.jetbrains.com/issue/KTOR-3001 is resolved
-@OptIn(InternalAPI::class)
 class KtorUtilsTest {
 
     @Test
