+++ runtime/protocol/http-client-engines/test-suite/build.gradle.kts
@@ -27,7 +27,7 @@ kotlin {
             dependencies {
                 implementation("io.ktor:ktor-server-cio:$ktorVersion")
 
-                implementation(project(":runtime:protocol:http-client-engines:http-client-engine-ktor"))
+                implementation(project(":runtime:protocol:http-client-engines:http-client-engine-default"))
                 implementation(project(":runtime:protocol:http-client-engines:http-client-engine-crt"))
             }
         }
