+++ runtime/protocol/http-client-engines/test-suite/jvm/src/aws/smithy/kotlin/runtime/http/test/util/AbstractEngineTestJVM.kt
@@ -6,8 +6,8 @@
 package aws.smithy.kotlin.runtime.http.test.util
 
 import aws.smithy.kotlin.runtime.http.engine.HttpClientEngine
+import aws.smithy.kotlin.runtime.http.engine.SdkHttpEngine
 import aws.smithy.kotlin.runtime.http.engine.crt.CrtHttpEngine
-import aws.smithy.kotlin.runtime.http.engine.ktor.KtorEngine
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.runBlocking
 import kotlinx.coroutines.withTimeout
@@ -16,7 +16,7 @@ import kotlin.time.Duration
 
 internal actual fun engines(): List<HttpClientEngine> =
     listOf(
-        KtorEngine(),
+        SdkHttpEngine(),
         CrtHttpEngine()
     )
